pyproject.toml
[project]
name = "WordleBot"
version = "0.1.0"
description = "基于 NoneBot2 编写的 Wordle 机器人"
authors = [
    {name = "zhaomaoniu", email = "2667292003@qq.com"},
]
dependencies = [
    "Pillow>=10.1.0",
    "nonebot2[httpx,websockets]>=2.1.2",
    "nonebot-plugin-alconna>=0.34.1",
    "nonebot-adapter-qq>=1.3.4",
    "httpx>=0.25.2",
]
requires-python = ">=3.8"
readme = "README.md"
license = {text = "MIT"}

